format/mainPage_treeInfos 카카오 공유 url, content 수정/#11
