--- src/components/common/Sharing/SharingModal/SharingInfo.js
@@ -1,9 +1,0 @@
-export const SharingInfo = {
-  title: '어쩔트리로 트리를 공유해보세요!',
-  baseUrl: 'https://63860b066ed8c30db7349d3b--comforting-moonbeam-e55e7d.netlify.app',
-  address: 'address',
-  addressTitle: 'addressTitle',
-  description: 'description',
-  imageUrl:
-    'http://k.kakaocdn.net/dn/bSbH9w/btqgegaEDfW/vD9KKV0hEintg6bZT4v4WK/kakaolink40_original.png',
-}
