+++ src/components/common/Sharing/SharingModal/index.jsx
@@ -2,16 +2,16 @@ import React, { useEffect } from 'react'
 import { CopyToClipboard } from 'react-copy-to-clipboard'
 import { useSelector } from 'react-redux'
 // import { shareToTwitter, shareToFacebook } from '../ShareLinks'
-import { SharingInfo } from './SharingInfo'
+import { ShareToKakaoTalk, SharingInfo } from '../ShareToKakaoTalk.jsx'
 import { Sharing_kakaoIcon, Sharing_urlIcon } from '../../../Icons'
 import * as S from './style'
 
 export const SharingModal = ({ setUrlCopiedAlert }) => {
-  const { tree_id } = useSelector((store) => store.tree.tree)
+  const { tree_id, tree_name, tree_addr } = useSelector((store) => store.tree.tree)
 
   useEffect(() => {
     const script = document.createElement('script')
-    script.src = 'https://developers.kakao.com/sdk/js/kakao.js' // 카카오톡 SDK
+    script.src = 'https://developers.kakao.com/sdk/js/kakao.js'
     script.async = true
 
     document.body.appendChild(script)
@@ -20,38 +20,6 @@ export const SharingModal = ({ setUrlCopiedAlert }) => {
       document.body.removeChild(script)
     }
   }, [])
-
-  const shareToKakaoTalk = () => {
-    if (window.Kakao === undefined) {
-      return
-    }
-
-    const kakao = window.Kakao
-
-    if (window.kakao) {
-      if (!kakao.isInitialized()) {
-        // 두번째 step 에서 가져온 javascript key 를 이용하여 initialize
-        kakao.init(process.env.REACT_APP_KAKAO_API_KEY)
-      }
-
-      kakao.Share.sendDefault({
-        objectType: 'location',
-        address: SharingInfo.address,
-        addressTitle: SharingInfo.addressTitle,
-        content: {
-          title: SharingInfo.title,
-          description: SharingInfo.description,
-          imageUrl: SharingInfo.imageUrl,
-          link: {
-            // [내 애플리케이션] > [플랫폼] 에서 등록한 사이트 도메인과 일치해야 함
-            mobileWebUrl:
-              'https://63860b066ed8c30db7349d3b--comforting-moonbeam-e55e7d.netlify.app',
-            webUrl: 'https://63860b066ed8c30db7349d3b--comforting-moonbeam-e55e7d.netlify.app',
-          },
-        },
-      })
-    }
-  }
 
   return (
     <S.Container>
@@ -62,7 +30,7 @@ export const SharingModal = ({ setUrlCopiedAlert }) => {
             <S.BtnTitle>URL 복사</S.BtnTitle>
           </S.BtnBox>
         </CopyToClipboard>
-        <S.BtnBox onClick={() => shareToKakaoTalk()}>
+        <S.BtnBox onClick={() => ShareToKakaoTalk(tree_addr, tree_name)}>
           <Sharing_kakaoIcon />
           <S.BtnTitle>카카오톡</S.BtnTitle>
         </S.BtnBox>
